Rewrote smartPlacement()
Instead of trying to find the position with the least overlaps, including all
stacked windows with the same weight, we now try to determine the place where
the lowest window on stack is visible, to put the new window there.

In effect, this means all windows hidden by another one are ignored, only
windows visible to the user should matter. We put the new window on top of the
window the lowest on stack, i.e. the last one to be activated/(fully) visible.